we back (wooo)
this is pretty much an interim commit. the entire game has been being revamped for way too long to be server-sided rather than client sided. i wont put too long of a changelog as i hope to have most, if not all of this done by the weekend.

what is working right now is:
- objects (tree's specifically, but the changes to the fishing areas and the teleporters SHOULD be quite simple to fix.
- inventory
- quests
- accessories (i believe)
- skills

the difference between this new build and the older ones is that, for instance, when you chop down a tree and it collapses, everyone will see the tree collapse, as the server is handling one instance of the game rather than each client having their own.